IOS Calculator App - TeamtTreehouseContest
This is a basic iPhone app. It explores Xcode and the objective-c
programming language, as well as having visual-design considerations.
Adobe Illustrator was used to manipulate an eps file. That file came
from Shutterstock and was also an exploration in using graphic-editing
software via having to modify for specific purposes. Finally, it should
be noted that the code here is modified version of a tutorial found at
Lynda.com.